Apollo: Fix media position transmission via RemoteControlClient
Using the legacy call setPlaybackState(int) resets the position to
"unknown", therefore the new API has to be used for PLAYSTATE_CHANGED
as well.

It's also necessary to update the media position when the track has
changed as it doesn't trigger PLAYSTATE_CHANGED or POSITION_CHANGED.
The drift check would eventually correct this, but for some seconds
the position would be wrong.

Change-Id: I39c85c7092c36d0bbc4817de7e72654524d8336c